requirements.txt
aws-cdk-lib==2.22.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-go-alpha==2.22.0a0
python-dotenv==0.20.0
